Add scheme to PoolKey and let port be None.
This fixes a panic that can happen on redirects to schemes unrecognized
by Url.

Remove the test-only pool.get() accessor. This was used in only test,
agent_pool in range.rs. This seemed like it was testing the agent more
than it was testing ranges, so I moved it to agent.rs and edited to
remove the range-testing parts.

Also, reject unrecognized URLs earlier in connect_socket so they don't
reach try_get_connection.