Split out RequestBuilder from Request.
This formalizes the builder pattern with regards to a Request. All
methods like `ureq::get` that formerly returned a Request now return a
RequestBuilder, so you can add headers and params the same as before.
Also, RequestBuilder implements `call()` and `send*()` as shortcuts to
immediately build and send a Request, so call chains don't get longer.

Add the ability to set a URL on a request directly, in addition to the
ability to pass an unparsed `&str`. Remove some of the individual field
getters from Request, like get_method, get_url, get_scheme, and get_host.

Remove the `qstring` dependency. We can achieve the same thing
using Url's query_pairs_mut.

Delete the `.proxy()` convenience method on Request.