Default to native-tls on platforms that need it
This adds config gates in Cargo.toml and in the source code that make
native-tls the default on platforms that don't support rustls.

To keep our matrix of features and platforms simple, I've eliminated the
`tls` feature. TLS is now always built-in, and which library gets used
depends on your platform.