Add Request::call_writer for writing request body using std::io::Write
Request.call_writer() sends the request headers, and returns a RequestStream
object that can be used to write the request body. This is more convenient
than trying to create a Read object that does the same thing.

This also makes some major changes to the middleware trait, to support the
call_writer API. It's now split into `handle_request` and `handle_response`.