add https proxy support
Adding HTTPS proxy support reuses `Stream` for the connection, so nested
TlsStreams can be used, one for the proxy and one for the connection to
the target host.

This comes at the cost of making Stream and parts of the interface
public. In addition, the `Sync` trait is required for Errors to continue
working (std::io::Error requirement).

As all TlsConnectors seem to be generic over the Stream type, replacing
the TcpStream type with Stream seems to not be an issue.

HTTPS proxy currently uses a default native_tls TlsConnector when the
'native-tls' feature is set, otherwise it falls back to
'default_tls_config()'.

Signed-off-by: Mira Limbeck <m.limbeck@proxmox.com>