gzip/brotli make sure wrapped stream reads to n=0
This commit adds two intermediary wrappers to the gzip/brotli decoding
cases, DidReadZero and CompressionRead.

The inner, DidReadZero, keeps track of whether the wrapped stream has
read to n=0. The wrapped stream will typically be a
`PoolReturnRead(ChunkedDecoder)` or `PoolReturnRead(LimitRead)`, both
which are fine to read multiple times to 0.

The outher CompressionRead will on n=0 unpack the inner DidReadZero
and ensure we do read to 0.