Fix connection reuse for empty response bodies
For empty response bodies, there's no guarantee that
`PoolReturnRead::read` will ever be called, so we might never notice
that the body is exhausted and can be returned to the pool. Solve this
by adding an exhaustion check to `PoolReturnRead::new`. Annoyingly, this
requires changing its API (adding a new trait bound + making `new`
fallible), but fortunately it's an internal API and the callers are easy
to update.

Fixes #555