fix #557: read proxy response until \r\n\r\n
Headers are terminated by CRLF according to the spec [0]. Between
headers and the body there's an additional CRLF. This means we have to
parse until we get 2 CRLFs right after each other to know when we can
stop reading from the stream.

Since the `Response` struct already handles this, we can reuse that
instead of custom stream reading and response parsing.

[0] https://www.ietf.org/rfc/rfc2068.txt (6 Response)

Signed-off-by: Mira Limbeck <m.limbeck@proxmox.com>