Enable all features in documentation and utilize `doc_auto_cfg`
Testing all features instead of just a specific set makes it possible
for the CI to spot various mistakes in currently-untested doc comments,
such as those behind the `http-interop` feature.  More importantly,
enabling _all features_ on docs.rs as well massively helps visibility
in available features and functionality to the end user.  And `doc_cfg`/
`doc_auto_cfg` automatically provide them the right information block
with a (list of) feature(s) they need to enable to get access to a
certain type or function.

Note that this requires a `nightly` feature and is hence behind a
`--cfg` flag that is only enabled when building for docs.rs (using
`nightly`). It can be reproduced locally via `RUSTDOCFLAGS=--cfg=docsrs
cargo +nightly doc`.

Additionally the CI script is updated to pass `-Dwarnings` to `rustdoc`
so that it can adequately fail the CI when incorrect documentation is
encountered.