ntls: Convert native_tls::WouldBlock to io::Error::TimedOut
When the tls handshake starts and the server doesn't respond, the
std::net::TcpStream will return WouldBlock or TimedOut when a
read/write times out to the socket, which in turn native_tls will
convert to a HandshakeError::WouldBlock. Instead of returning back a
io::Error with no type, return io::Error with kind TimedOut.