fix: use reader for bytes http response transformation
The From implementation to transform a ureq::Response into an
http::Response<Vec<u8>> is using the into_string() method. This method
only allows for UTF-8 characters in the body payload, thus the
result is inconsistent for non UTF-8 characters like protobuffers (binary).

The proposed solution uses the reader instead of transforming the input
into a UTF-8 character.