Expand documented edge cases
## Problem

I was under the mistaken impression that the reader returned by the `into_reader` didn't perform checks for Content-Length. It does, but even after consulting the documentation in several places, I moved forward with adding my own check in https://github.com/heroku/libcnb.rs/pull/947. A co-worker corrected my implementation, which led me to discover it wasn't needed. Neither of us came to the conclusion that we were already protected via documentation alone.

One of the doc locations that I checked was `into_reader`.

## Fix

I added two error cases to the docs where the reader can (and should) make errors: when a connection is closed before Content-Length bytes are received and if an ungraceful close of the TCP connection occurs.

At the same time, the subject of the bullet point on the reader not being "limited" wasn't clear to me. To help, I added some more details.

I understand the documents weren't written for me personally, and we cannot add every possible caveat to every possible source of documentation. Feel free to take this feedback and run with it if you like. Thanks for this library!