Track size for files
To allow producing Content-Length headers when using a File as the body,
try to get the size of the file before type-erasing it. If the File
points to something other than a regular file (e.g. a named pipe), or if
we fail to get the metadata for some reason, use the current behavior of
sending it with Content-Transfer-Encoding: chunked.

Issue: #1098